pulled activedata out of cm_dossier
